requirements.txt
requests
ipwhois
traceroute
python-whois
